sandbox: use channel for worker loop quit signal
Instead of the current implementation that uses an `atomic.Bool` rather awkwardly,
use a channel instead (close to signal worker loop exit), which is more idiomatic.

Note that this is also in preparation for a larger ongoing refactoring effort,
where we will fundamentally rework the worker loop and driver shutdown logic.