docker: refactor driver worker loop implementation
This change performs a large refactoring of the worker loop implementation
used for starting guest containers and maintaining the "ready" queue.

Notably, we prepare for supporting multiple concurrent worker goroutines.

We will later also refactor the `Driver.Close` method to change how (and
which) resources are cleaned up upon a graceful shutdown of the driver.