docker: refactor container destroy/remove behavior
This change updates the `Driver.Destroy` and `Driver.remove` methods to
only try to remove a specified guest container once and then forget it.

This is part of a larger ongoing refactoring effort where we change the
way we implement guest container removal when `Driver.Close` is called.

The basic idea is that, in the future, the `Driver.Close` method will
only remove already started and not yet acquired guest containers (i.e.,
exactly those containers in the "ready" queue); and we will make it the
responsibility of callers to the `Driver.Acquire` method to clean up any
guest containers already acquired by themselves when no longer required.

Note that we keep the existing structure of the `Driver.remove` method,
specifically the seemingly redundant error check, because we will later
add a debug logging statement here and expand the usage of the context.